refactor(resource): move all resource services into a single folder to make the folder structures more readable
ref: #371
