+++ lib/src/core/dataloaders/kanji_dataloader.dart
@@ -1,7 +1,7 @@
 import "package:flutter/foundation.dart";
 import "package:kana_to_kanji/src/core/dataloaders/resource_dataloader.dart";
 import "package:kana_to_kanji/src/core/models/resources/kanji.dart";
-import "package:kana_to_kanji/src/core/services/kanji_service.dart";
+import "package:kana_to_kanji/src/core/services/resources/kanji_service.dart";
 import "package:kana_to_kanji/src/core/utils/kana_utils.dart";
 
 class KanjiDataLoader extends ResourceDataLoader<Kanji> {
