+++ lib/src/core/repositories/resource_repository.dart
@@ -1,6 +1,6 @@
 import "package:flutter/foundation.dart";
 import "package:kana_to_kanji/src/core/models/resources/resource.dart";
-import "package:kana_to_kanji/src/core/services/resource_data_service.dart";
+import "package:kana_to_kanji/src/core/services/resources/resource_data_service.dart";
 import "package:stacked/stacked.dart";
 
 abstract class ResourceRepository<
