+++ lib/src/core/services/resources/kanji_service.dart
@@ -4,7 +4,7 @@ import "package:kana_to_kanji/src/core/dataloaders/kanji_dataloader.dart";
 import "package:kana_to_kanji/src/core/models/resources/resources.dart"
     show Kanji, ResourceUid;
 import "package:kana_to_kanji/src/core/services/database_service.dart";
-import "package:kana_to_kanji/src/core/services/resource_data_service.dart";
+import "package:kana_to_kanji/src/core/services/resources/resource_data_service.dart";
 import "package:kana_to_kanji/src/locator.dart";
 import "package:sqflite/sqflite.dart";
