+++ lib/src/core/services/resources/resource_data_service.dart
