+++ test/src/core/services/resources/group_service_test.dart
@@ -5,14 +5,14 @@ import "package:kana_to_kanji/src/core/models/paginated_response.dart";
 import "package:kana_to_kanji/src/core/models/resources/group.dart";
 import "package:kana_to_kanji/src/core/models/resources/resource_uid.dart";
 import "package:kana_to_kanji/src/core/services/database_service.dart";
-import "package:kana_to_kanji/src/core/services/group_service.dart";
+import "package:kana_to_kanji/src/core/services/resources/group_service.dart";
 import "package:kana_to_kanji/src/locator.dart";
 import "package:logger/logger.dart";
 import "package:mockito/annotations.dart";
 import "package:mockito/mockito.dart";
 
-import "../../../dummies/group.dart";
-import "../../../helpers.dart";
+import "../../../../dummies/group.dart";
+import "../../../../helpers.dart";
 @GenerateNiceMocks([MockSpec<ResourceDataLoader>(), MockSpec<Logger>()])
 import "group_service_test.mocks.dart";
