+++ test/src/core/services/resources/kana_service_test.dart
@@ -4,16 +4,16 @@ import "package:kana_to_kanji/src/core/dataloaders/resource_dataloader.dart";
 import "package:kana_to_kanji/src/core/models/paginated_response.dart";
 import "package:kana_to_kanji/src/core/models/resources/resources.dart";
 import "package:kana_to_kanji/src/core/services/database_service.dart";
-import "package:kana_to_kanji/src/core/services/group_service.dart";
-import "package:kana_to_kanji/src/core/services/kana_service.dart";
+import "package:kana_to_kanji/src/core/services/resources/group_service.dart";
+import "package:kana_to_kanji/src/core/services/resources/kana_service.dart";
 import "package:kana_to_kanji/src/locator.dart";
 import "package:logger/logger.dart";
 import "package:mockito/annotations.dart";
 import "package:mockito/mockito.dart";
 import "package:sqflite/sqflite.dart";
 
-import "../../../dummies/dummies.dart";
-import "../../../helpers.dart";
+import "../../../../dummies/dummies.dart";
+import "../../../../helpers.dart";
 
 @GenerateNiceMocks([MockSpec<ResourceDataLoader>(), MockSpec<Logger>()])
 import "kana_service_test.mocks.dart";
