+++ test/src/core/services/resources/kanji_service_test.dart
@@ -4,9 +4,10 @@ import "package:kana_to_kanji/src/core/models/paginated_response.dart";
 import "package:kana_to_kanji/src/core/models/resources/kanji.dart";
 import "package:kana_to_kanji/src/core/models/resources/resource_uid.dart";
 import "package:kana_to_kanji/src/core/services/database_service.dart";
-import "package:kana_to_kanji/src/core/services/group_service.dart" as groups;
-import "package:kana_to_kanji/src/core/services/kanji_service.dart";
-import "package:kana_to_kanji/src/core/services/vocabulary_service.dart"
+import "package:kana_to_kanji/src/core/services/resources/group_service.dart"
+    as groups;
+import "package:kana_to_kanji/src/core/services/resources/kanji_service.dart";
+import "package:kana_to_kanji/src/core/services/resources/vocabulary_service.dart"
     as vocab;
 import "package:kana_to_kanji/src/locator.dart";
 import "package:logger/logger.dart";
@@ -15,8 +16,8 @@ import "package:mockito/mockito.dart";
 import "package:sqflite/sqflite.dart";
 import "package:sqflite/utils/utils.dart";
 
-import "../../../dummies/dummies.dart";
-import "../../../helpers.dart";
+import "../../../../dummies/dummies.dart";
+import "../../../../helpers.dart";
 @GenerateNiceMocks([MockSpec<KanjiDataLoader>(), MockSpec<Logger>()])
 import "kanji_service_test.mocks.dart";
